copy_gh_file_to_bucket/scripts/requirements.txt
GitPython
google-cloud-storage